Fix test runner

+++ test/runner.py
@@ -151,7 +151,10 @@ def test_normal_sprites(sprites = None):
             lrdir = rdir.lower()
             if 'routines' in lrdir:
                 for file in glob.glob(rdir + '*'):
-                    copyfile(file, 'pixi/routines/' + os.path.basename(file))
+                    if os.path.isfile(file):
+                        copyfile(file, 'pixi/routines/' + os.path.basename(file))
+                    elif os.path.isdir(file):
+                        copytree(file, 'pixi/routines/' + os.path.basename(file), dirs_exist_ok=True)
         copyfile('_header.asm', 'pixi/sprites/_header.asm')
         os.remove('pixi/sprites/list' + folder.split(os.sep)[-1] + '.txt')
         copyfile('pixi/base.smc', 'pixi/work.smc')
